Conflicts:
	Applications/mitkDiffusion/CMakeLists.txt
	Applications/mitkWorkbench/CMakeLists.txt
	CMake/mitkMacroCreateModule.cmake
	CMake/mitkMacroCreateModuleTests.cmake
	CMake/mitkMacroUseModule.cmake
	Core/Code/DataManagement/mitkTimeSlicedGeometry.cpp
	Examples/Overlays/CMakeLists.txt
	Examples/Tutorial/CMakeLists.txt
	Modules/CMakeLists.txt
	Modules/DicomUI/CMakeLists.txt
	Modules/IGT/IGTTutorial/CMakeLists.txt
	Modules/OpenCVVideoSupport/UI/CMakeLists.txt
	Modules/Overlays/CMakeLists.txt
	Modules/Qmitk/CMakeLists.txt
	Modules/QmitkExt/files.cmake
	README.md

+++ Modules/Segmentation/Interactions/mitkBinaryThresholdTool.cpp
@@ -190,7 +190,7 @@ void mitk::BinaryThresholdTool::SetupPreviewNodeFor( DataNode* nodeForThresholdi
       }
       else
       {
-        m_CurrentThresholdValue = (m_SensibleMaximumThresholdValue + m_SensibleMinimumThresholdValue)/2;
+        m_CurrentThresholdValue = (m_SensibleMaximumThresholdValue + m_SensibleMinimumThresholdValue) / 2.0;
       }
 
       IntervalBordersChanged.Send(m_SensibleMinimumThresholdValue, m_SensibleMaximumThresholdValue, m_IsFloatImage);
@@ -317,8 +317,8 @@ void mitk::BinaryThresholdTool::OnRoiDataChanged()
 
     tmpNode->SetData(tmpImage);
 
-    m_SensibleMaximumThresholdValue = static_cast<int> (roiFilter->GetMaxValue());
-    m_SensibleMinimumThresholdValue = static_cast<int> (roiFilter->GetMinValue());
+    m_SensibleMaximumThresholdValue = static_cast<double> (roiFilter->GetMaxValue());
+    m_SensibleMinimumThresholdValue = static_cast<double> (roiFilter->GetMinValue());
     SetupPreviewNodeFor( tmpNode );
     m_NodeForThresholding = tmpNode;
